+++ msg/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class MsgConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "msg"
+    verbose_name = "Messages"