+++ msg/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from . import views
+
+app_name = "msg"
+
+urlpatterns = [
+    path("", views.send, name="send"),
+]